+++ activerecord/CHANGELOG.md
@@ -1,3 +1,19 @@
+*   Make `assert_queries` and `assert_no_queries` assertions public.
+
+    To assert the expected number of queries are made, Rails internally uses
+    `assert_queries` and `assert_no_queries`. These assertions can be now
+    be used in applications as well.
+
+    ```ruby
+    class ArticleTest < ActiveSupport::TestCase
+      test "queries are made" do
+         assert_queries(1) { Article.first }
+      end
+    end
+    ```
+
+    *Petrik de Heus*
+
 *   Fix `has_secure_token` calls the setter method on initialize.
 
     *Abeid Ahmed*
